refactor: Log 패키지 위치 변경
- 도메인 코드와 test코드를 domain/log 하위로 이동했습니다.
