+++ logbat/src/test/java/info/logbat/domain/project/domain/log/domain/values/LogDataTest.java
@@ -1,10 +1,12 @@
-package info.logbat.log.domain.values;
+package info.logbat.domain.project.domain.log.domain.values;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import info.logbat.domain.log.domain.values.LogData;
 
 @DisplayName("LogData VO 테스트")
 class LogDataTest {
